Generate interactive map of election results

Requires data from wrangle_results() function joined with a shapefile. See maps vignette for more info.

map_election_results(
  joined_df,
  win_col,
  lose_col,
  election_district_col,
  the_winner_palette = "Purples",
  the_loser_palette = "Greens",
  map_tiles = "CartoDB.Positron"
)

Arguments

joined_df
sf object shapefile of election districts that's been joined with election result data.table generated by wrangle_results() function

win_col
character string name of column with winner's data by Pct or Margin

lose_col
character string name of column with loser's data by Pct or Margin

election_district_col
character string name of column with election district data - should be name of 1st column from the wrangle_results() data frame

the_winner_palette
character string of colors for leaflet colorNumeric function

the_loser_palette
character string of colors for leaflet colorNumeric function

map_tiles
character string for available leaflet tile providers, defaults to "CartoDB.Positron". See http://leaflet-extras.github.io/leaflet-providers/preview/ for options.

Value

leaflet map object