Generate interactive map of turnout

Requires data with election district and turnout percent as a decimal between 0 and 1 joined with a shapefile.

map_turnout(
  joined_df,
  turnout_col,
  election_district_col,
  the_palette = "Greens",
  map_tiles = "CartoDB.Positron"
)

Arguments

joined_df
sf object shapefile of election districts that's been joined with election result

turnout_col
character string name of column with turnout data

election_district_col
character string name of column with election district data - should be name of 1st column from the wrangle_results() data frame

the_palette
character string for color for leaflet colorNumeric function

map_tiles
character string for available leaflet tile providers, defaults to "CartoDB.Positron". See http://leaflet-extras.github.io/leaflet-providers/preview/ for options.

Value

leaflet map object