Create a quick searchable, sortable table of results with the DT package

Features colored bars in percent columns and allows for searching by regular expression

quick_table(
  election_df,
  pagination = 20,
  win_color = "#c2a5cf",
  lose_color = "#a6dba0",
  use_regex_searching = TRUE,
  n_turnout_cols = 0
)

Arguments

election_df
data table of election results generated by wrangle_results() including Total row

pagination
integer how many rows to display on each color

win_color
character string for color bar in winner's percent column, defaults to "#c2a5cf" (a light purple). Use "none" for no colored bar.

lose_color
character string for color bar in winner's percent column, defaults to "#a6dba0" (a light green). Use "none" for no colored bar.

use_regex_searching
logical whether to allow searching by regular expression, defaults to TRUE

n_turnout_cols
integer how many turnout columns have been joined to election_df: 0 for no data about turnout; 1 for a column with turnout percent only; or 3 for columns containing total votes cast, total registered voters, and percent turnout. Defaults to 0.

Value

html widget with interactive table of results

Examples

myfile <- system.file("extdata", "FakeElectionResults.xlsx", package = "elections2")
my_election_results <- wrangle_results(myfile)
quick_table(my_election_results)
quick_table(my_election_results, win_color = "none", lose_color = "none")