Generate table of results with more than 2 columns, including sparkline bar graphs for percent data

Generate table of results with more than 2 columns, including sparkline bar graphs for percent data

quick_table_more_cols(
  election_df,
  graph_cols,
  show_pcts = FALSE,
  pagination = 20,
  show_sparklines = TRUE,
  use_regex_searching = TRUE,
  my_palette = c("none")
)

Arguments

election_df
data table of election results generated by wrangle_more_cols() function

graph_cols
vector of character strings with names of columns containing original vote or percent results (not totals or margins)

show_pcts
logical whether election_df data is showing percents as opposed to vote totals. Defaults to FALSE.

pagination
integer number of rows to display per page, defaults to 20

show_sparklines
logical if you want table to include a column with mini bar charts. Defaults to TRUE.

my_palette
character string with colors to display in Winner and RunnerUp columns. Defaults to "none" for no color.

Value

Examples

myfile <- system.file("extdata", "Fake3Columns.xlsx", package = "elections2")
mydata <- wrangle_more_cols(myfile, c("Red", "Blue", "Orange", "Green", "Purple"), show_pcts = TRUE)
quick_table_more_cols(mydata, c("Red", "Blue", "Orange", "Green", "Purple"), show_pcts = TRUE, my_palette = c("#DF536B", "#2297E6", "orange", "green", "purple"))