Get winner for race with three or more candidates

Get winner for race with three or more candidates

Usage

wrangle_more_cols(
  election_results_file,
  votes_cols,
  show_pcts = FALSE,
  show_runnerup = TRUE,
  show_margin = TRUE,
  show_margin_2vs3 = FALSE,
  use_column_numbers = FALSE
)

Arguments

election_results_file

character string with name of Excel or CSV file in format with election district in 1st col. You can also use the name of an R data frame if you already have your results data in R, but the object name MUST be in quotation marks.

votes_cols

vector of character strings with names of columns containing votes (or numeric vector if use_column_numbers is TRUE)

show_pcts

logical TRUE to display results as percents. Defaults to FALSE for vote totals.

show_runnerup

logical TRUE to show 2nd place finisher. Defaults to TRUE.

show_margin

logical TRUE to show margin between 1st and 2nd place. Defaults to TRUE.

show_margin_2vs3

logical TRUE to show margin between 2nd and 3rd place. Defaults to FALSE

use_column_numbers

logical TRUE to use column index numbers instead of character strings with column names in votes_cols. Defaults to FALSE.

Value

data table with results, totals, and district winner by number of votes

Examples

myfile <- system.file("extdata", "Fake3Columns.xlsx", package = "elections2")
mydata <- wrangle_more_cols(myfile, c("Red", "Blue", "Green"), show_pcts = TRUE, show_runnerup = FALSE)
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames
#> Warning: cannot xtfrm data frames